# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Igloo",
    "description": "The Igloo control plane API.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Gets a job.",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sandboxes": {
      "get": {
        "tags": [
          "sandboxes"
        ],
        "summary": "Lists sandboxes, ordered by id, optionally filtered by labels (`label=key=value`, repeatable).",
        "operationId": "listSandboxes",
        "parameters": [
          {
            "name": "label",
            "in": "query",
            "description": "`key=value`; every label must match",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "`next_cursor` of the previous page",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1 to 200; 50 when omitted",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxList"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sandboxes"
        ],
        "summary": "Creates a sandbox from a registered snapshot.",
        "operationId": "createSandbox",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Makes retries safe",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxResource"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes/{id}": {
      "get": {
        "tags": [
          "sandboxes"
        ],
        "summary": "Gets a sandbox.",
        "operationId": "getSandbox",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes/{id}/exec": {
      "post": {
        "tags": [
          "sandboxes"
        ],
        "summary": "Runs a process in a sandbox as a job.",
        "operationId": "execInSandbox",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sandboxes/{id}/stop": {
      "post": {
        "tags": [
          "sandboxes"
        ],
        "summary": "Stops a sandbox. Idempotent.",
        "operationId": "stopSandbox",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxResource"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSandboxRequest": {
        "type": "object",
        "description": "Creates a sandbox from a snapshot.",
        "required": [
          "snapshot"
        ],
        "properties": {
          "env": {
            "type": "object",
            "description": "Environment variables of every process.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "isolation": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Isolation",
                "description": "Separation from the host; `any` when omitted."
              },
              {
                "type": "null"
              }
            ]
          },
          "labels": {
            "type": "object",
            "description": "Metadata for grouping and filtering.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "limits": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Limits",
                "description": "CPU and memory bounds; secure defaults when omitted."
              },
              {
                "type": "null"
              }
            ]
          },
          "network": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/NetworkPolicy",
                "description": "Network access; `deny_all` when omitted."
              },
              {
                "type": "null"
              }
            ]
          },
          "repo": {
            "type": [
              "string",
              "null"
            ],
            "description": "The repository the sandbox works on (`repo_...`); its secrets become available to jobs."
          },
          "snapshot": {
            "type": "string",
            "description": "The snapshot to start from."
          }
        }
      },
...
      "DesiredState": {
        "type": "string",
        "description": "Whether the sandbox should run.",
        "enum": [
          "running",
          "stopped"
        ]
      },
...
      "ExecRequest": {
        "type": "object",
        "description": "Runs a process in a sandbox.",
        "required": [
          "argv"
        ],
        "properties": {
          "argv": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The program and its arguments."
          },
          "env": {
            "type": "object",
            "description": "Extra environment, layered over the sandbox's.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "secrets": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Secrets of the sandbox's repository to expose as environment variables, by name."
          },
          "timeout_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "How long the process may run, 1 to 86400 seconds; one hour when omitted.",
            "minimum": 0
          }
        }
      },
      "FieldError": {
        "type": "object",
        "description": "One invalid field of a request.",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "The field path, such as `\"limits.millicpus\"`."
          },
          "message": {
            "type": "string",
            "description": "Why it is invalid."
          }
        }
      },
...
      "Isolation": {
        "type": "string",
        "description": "How strongly a sandbox is separated from its host.",
        "enum": [
          "any",
          "container"
        ]
      },
      "JobPhase": {
        "type": "string",
        "description": "Where a job is in its lifecycle.",
        "enum": [
          "queued",
          "leased",
          "running",
          "finished",
          "failed",
          "cancelled"
        ]
      },
      "JobResource": {
        "type": "object",
        "description": "A job.",
        "required": [
          "id",
          "sandbox",
          "argv",
          "env",
          "timeout_seconds",
          "phase",
          "submitted_at"
        ],
        "properties": {
          "argv": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The program and its arguments."
          },
          "env": {
            "type": "object",
            "description": "Extra environment.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "exit_code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The exit code, when finished."
          },
          "failure_reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why it failed, when failed."
          },
          "id": {
            "type": "string",
            "description": "Its id (`job_...`)."
          },
          "phase": {
            "$ref": "#/components/schemas/JobPhase",
            "description": "The phase."
          },
          "sandbox": {
            "type": "string",
            "description": "The sandbox it runs in."
          },
          "secrets": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The repository secrets it is given, by name."
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "description": "When it was submitted."
          },
          "timeout_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "How long the process may run.",
            "minimum": 0
          }
        }
      },
...
      "Limits": {
        "type": "object",
        "description": "CPU and memory bounds.",
        "required": [
          "millicpus",
          "memory_mib"
        ],
        "properties": {
          "memory_mib": {
            "type": "integer",
            "format": "int32",
            "description": "Memory in MiB, 128 to 262144.",
            "minimum": 0
          },
          "millicpus": {
            "type": "integer",
            "format": "int32",
            "description": "CPU in thousandths of a core, 100 to 64000.",
            "minimum": 0
          }
        }
      },
...
      "NetworkPolicy": {
        "type": "string",
        "description": "Network access of a sandbox.",
        "enum": [
          "deny_all",
          "allow_all"
        ]
      },
...
      "Problem": {
        "type": "object",
        "description": "An error response. `code` is stable and machine-readable; `title` and `detail` are for\nhumans.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "The stable error code, such as `\"sandbox.not_found\"`."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "What went wrong in this occurrence."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Every invalid field, for validation problems."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "A short summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "A URI identifying the problem type."
          }
        }
      },
...
      "SandboxList": {
        "type": "object",
        "description": "A page of sandboxes.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SandboxResource"
            },
            "description": "The sandboxes, ordered by id."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass as `cursor` to get the next page; absent on the last page."
          }
        }
      },
      "SandboxPhase": {
        "type": "string",
        "description": "Where a sandbox is in its lifecycle.",
        "enum": [
          "pending",
          "scheduled",
          "starting",
          "running",
          "stopping",
          "stopped",
          "failed"
        ]
      },
      "SandboxResource": {
        "type": "object",
        "description": "A sandbox: its desired spec and observed status.",
        "required": [
          "id",
          "snapshot",
          "desired",
          "limits",
          "network",
          "isolation",
          "env",
          "labels",
          "phase",
          "generation",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When it was created."
          },
          "desired": {
            "$ref": "#/components/schemas/DesiredState",
            "description": "Whether it should run."
          },
          "env": {
            "type": "object",
            "description": "Environment variables.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "failure_reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why it failed, when `phase` is `failed`."
          },
          "generation": {
            "type": "integer",
            "format": "int64",
            "description": "The spec generation; bumped by every spec change.",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "description": "Its id (`sbx_...`)."
          },
          "isolation": {
            "$ref": "#/components/schemas/Isolation",
            "description": "Separation from the host."
          },
          "labels": {
            "type": "object",
            "description": "Labels.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "limits": {
            "$ref": "#/components/schemas/Limits",
            "description": "CPU and memory bounds."
          },
          "network": {
            "$ref": "#/components/schemas/NetworkPolicy",
            "description": "Network access."
          },
          "observed_generation": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The generation the worker last reported acting on.",
            "minimum": 0
          },
          "phase": {
            "$ref": "#/components/schemas/SandboxPhase",
            "description": "The observed phase."
          },
          "repo": {
            "type": [
              "string",
              "null"
            ],
            "description": "The repository it works on."
          },
          "snapshot": {
            "type": "string",
            "description": "The snapshot it starts from."
          },
          "worker": {
            "type": [
              "string",
              "null"
            ],
            "description": "The worker it is assigned to."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}